refactor: add usererrorbutton hook
