+++ src/features/error/error-button.tsx
@@ -1,22 +1,9 @@
-import { useState } from 'react';
-
-import apiRoot from '@/api/api';
 import Button from '@/components/button';
 
-function ErrorButton() {
-  const [error, setError] = useState<Error | null>(null);
-  const [hasError, setHasError] = useState(false);
+import useErrorButton from './use-error-button';
 
-  const handleClick = async () => {
-    try {
-      await apiRoot().error();
-    } catch (error) {
-      if (error instanceof Error) {
-        setError(error);
-        setHasError(true);
-      }
-    }
-  };
+function ErrorButton() {
+  const { handleClick, error, hasError } = useErrorButton();
 
   if (hasError && error) {
     throw error;
